cloud-architecture/cloud-run-specialist/pyproject.toml
[project]
name = "cloudrun-agent"
version = "0.1.0"
description = "Claude Code agent for analyzing Google Cloud Run instances — resource waste, performance, pricing, traffic, and security"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = []
keywords = ["claude-code", "cloud-run", "gcp", "google-cloud", "devops", "infrastructure", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/thisisfatih/eng-team"
Issues = "https://github.com/thisisfatih/eng-team/issues"

[project.scripts]
cloudrun-agent = "cloudrun_agent.install:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cloudrun_agent"]


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
